refactor(identity-hub): move hub-specific components to src/ui/identity/
Applies the revised placement rule: `src/ui/components/` now hosts only
widgets with a plausible second consumer outside the Identities hub.
Hub-specific widgets move next to the tab modules that consume them.

Moved (git mv) into `src/ui/identity/`:

- identity_hub_tab_bar
- identity_hero_card
- onboarding_checklist
- identity_picker_card
- identity_picker_add_card
- identity_pill
- social_profile_gate_card
- request_card
- contact_row
- activity_row

`breadcrumb_pill` stays in `components/` — it is a generic label + icon
+ chevron pill usable anywhere a breadcrumb exists.

Import sites updated from `crate::ui::components::X` to the matching
`super::X` / `crate::ui::identity::X` path. `components/README.md`
rewritten to document the new placement rule; a new
`identity/README.md` catalogs the hub-local widgets.

No behavioural change — pure module relocation.
