+++ src/ui/components/mod.rs
@@ -1,32 +1,22 @@
-pub mod activity_row;
 pub mod address_input;
 pub mod amount_input;
 pub mod breadcrumb_pill;
 pub mod component_trait;
 pub mod confirmation_dialog;
-pub mod contact_row;
 pub mod contract_chooser_panel;
 pub mod dashpay_subscreen_chooser_panel;
 pub mod dpns_subscreen_chooser_panel;
 pub mod entropy_grid;
-pub mod identity_hero_card;
-pub mod identity_hub_tab_bar;
-pub mod identity_picker_add_card;
-pub mod identity_picker_card;
-pub mod identity_pill;
 pub mod identity_selector;
 pub mod info_popup;
 pub mod left_panel;
 pub mod left_wallet_panel;
 pub mod message_banner;
-pub mod onboarding_checklist;
 pub mod passphrase_modal;
 pub mod password_input;
 pub mod progress_overlay;
-pub mod request_card;
 pub mod secret_prompt_host;
 pub mod selection_dialog;
-pub mod social_profile_gate_card;
 pub mod styled;
 pub mod tokens_subscreen_chooser_panel;
 pub mod tools_subscreen_chooser_panel;
