+++ src/ui/identity/onboarding_checklist.rs
