+++ src/ui/identity/social_profile_gate_card.rs
